+++ tidy3d/plugins/mode/web.py
@@ -1,4 +1,5 @@
 """Web API for mode solver"""
-from ...web.api.mode import run
 
-__all__ = ["run"]
+from ...web.api.mode import run, run_batch
+
+__all__ = ["run", "run_batch"]